move analytics to utils
